+++ src/main/java/com/global/api/entities/Customer.java
@@ -36,8 +36,10 @@ public class Customer extends RecurringEntity<Customer> {
     private String department;
     private String status;
     private PhoneNumber phone;
+    private String language;
     private List<CustomerDocument> documents;
     private List<RecurringPaymentMethod> paymentMethods;
+    private Boolean isShippingAddressSameAsBilling;
 
     public Customer() {
         //super(Customer.class);